Ejercicios

+++ EJ6_DistanciaPuntos.py
@@ -1,10 +1,10 @@
 #EJ 6
 import math
 
-AX = int(input("Ingrese el primer valor de A: "))
-AY = int(input("Ingrese el segundo valor de A: "))
-BX = int(input("Ingrese el primer valor de B: "))
-BY = int(input("Ingrese el segundo valor de B: "))
+AX = float(input("Ingrese el primer valor de A: "))
+AY = float(input("Ingrese el segundo valor de A: "))
+BX = float(input("Ingrese el primer valor de B: "))
+BY = float(input("Ingrese el segundo valor de B: "))
 
 dist = math.sqrt((AX-BX)**2+(AY-BY)**2)
